bug: can not concat matrices that have an "ans" expression functions with more than 2 arguments do not work userFunctions.c: checking function name wasn't working correctly, made it its own function

+++ operator.h
@@ -13,7 +13,7 @@ matrix *matrixTwoArg(matrix *a, matrix *b, operatorStruct *ch, error_return *err
 error_return findFunction(char *buffer, numberStack *num, operatorStack *ch, vari *var, int8_t *tok, uint16_t *iterator, char *input);
 error_return findOperator(char *buffer, numberStack *num, operatorStack *ch, vari *var, int8_t *tok);
 
-char **separateMatrix(char *input, int length, error_return *error);
+char **separateMatrix(char *input, uint16_t length, error_return *error);
 matrix *extractMatrix(vari *var, uint16_t *iterator, char *input, error_return *error);
 
 void helpPrint(void);
